Agrega Error.h para evitar multiples definiciones
